@@ -1,9 +1,13 @@
 package io.manebot.plugin.music.source;
 
 import io.manebot.plugin.audio.mixer.input.AudioProvider;
+import io.manebot.plugin.audio.mixer.output.*;
+import io.manebot.plugin.audio.resample.*;
+import io.manebot.plugin.music.config.*;
 
 import java.io.*;
 
+import java.net.*;
 import java.util.Collections;
 
 import java.util.Map;
@@ -12,31 +16,32 @@ public interface AudioProtocol {
 
     /**
      * Opens an audio provider from the specified URL.
-     * @param url URL to open.
+     * @param uri URI to open.
      * @param format format name of the input stream (typically, the file extension of the container format).
      *               Can be null.
      * @param headers HTTP headers to use when opening the URL.
+     * @param bufferSize buffer size to use when opening the URL.
      * @return AudioProvider instance.
      */
-    AudioProvider open(String url, String format, Map<String, String> headers) throws IOException;
+    AudioProvider openProvider(URI uri, String format, Map<String, String> headers, int bufferSize) throws IOException;
 
     /**
      * Opens an audio provider from the specified URL.
-     * @param url URL to open.
+     * @param uri URI to open.
      * @param format format name of the input stream (typically, the file extension of the container format).
      * @return AudioProvider instance.
      */
-    default AudioProvider open(String url, String format) throws IOException {
-        return open(url, format, Collections.emptyMap());
+    default AudioProvider openProvider(URI uri, String format, int bufferSize) throws IOException {
+        return openProvider(uri, format, Collections.emptyMap(), bufferSize);
     }
 
     /**
      * Opens an audio provider from the specified URL, guessing the format as necessary.
-     * @param url URL to open.
+     * @param uri URI to open.
      * @return AudioProvider instance.
      */
-    default AudioProvider open(String url) throws IOException {
-        return open(url, null);
+    default AudioProvider openProvider(URI uri, int bufferSize) throws IOException {
+        return openProvider(uri, null, bufferSize);
     }
 
     /**
@@ -45,24 +50,35 @@ public interface AudioProtocol {
      * @param format format name of the input stream (typically, the file extension of the container format).
      * @return AudioProvider instance.
      */
-    AudioProvider open(InputStream inputStream, String format) throws IOException;
+    AudioProvider openProvider(InputStream inputStream, String format, int bufferSize) throws IOException;
 
     /**
      * Opens an audio provider from the specified input stream, guessing the format as necessary.
      * @param inputStream input stream of encoded samples to read from.
+     * @param bufferSize buffer size to use, in bytes.
      * @return AudioProvider instance
      */
-    default AudioProvider open(InputStream inputStream) throws IOException {
-        return open(inputStream, null);
+    default AudioProvider openProvider(InputStream inputStream, int bufferSize) throws IOException {
+        return openProvider(inputStream, null, bufferSize);
     }
+    
+    /**
+     * Opens an audio provider from the specified input stream, guessing the format and buffer size as necessary.
+     * @param inputStream input stream of encoded samples to read from.
+     * @return AudioProvider instance
+     */
+    AudioProvider openProvider(InputStream inputStream) throws IOException;
 
     /**
      * Opens an audio provider from the specified container file, guessing the format as necessary.
      * @param file container file to read from.
      * @return AudioProvider instance
      */
-    default AudioProvider open(File file) throws FileNotFoundException, IOException {
-        return open(new FileInputStream(file));
+    default AudioProvider openProvider(File file, int bufferSize) throws FileNotFoundException, IOException {
+        return openProvider(new FileInputStream(file), bufferSize);
     }
-
+    
+    AudioConsumer openConsumer(OutputStream outputStream, AudioDownloadFormat format) throws IOException;
+    
+    Resampler openResampler(AudioFormat input, AudioFormat output, int bufferSize) throws IOException;
 }
