@@ -1,8 +1,7 @@
 package io.manebot.plugin.music.source;
 
-import io.manebot.lambda.ThrowingFunction;
 import io.manebot.plugin.audio.mixer.input.AudioProvider;
-import io.manebot.plugin.music.Music;
+import io.manebot.plugin.music.*;
 import io.manebot.plugin.music.database.model.Community;
 import io.manebot.plugin.music.database.model.Track;
 
@@ -10,7 +9,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 import java.util.Objects;
-import java.util.function.Function;
 import java.util.stream.Stream;
 
 public class DatabaseTrackSource implements TrackSource {
@@ -31,42 +29,19 @@ public class DatabaseTrackSource implements TrackSource {
     }
 
     @Override
-    public Result find(Community community, URL url) throws IOException, IllegalArgumentException {
+    public Result find(Community community, URL url) {
         return Stream.of(community)
                 .map(c -> c.getTrack(url))
                 .filter(Objects::nonNull)
-                .map(track -> new LocalResult(
-                        track,
-                        ResultPriority.IDEAL
-                ))
+                .map(track -> new LocalResult(track, ResultPriority.IDEAL))
                 .findFirst().orElse(null);
     }
 
-    private class LocalResult extends Result {
-        public LocalResult(Track track, ResultPriority priority) {
-            this(track, priority,
-                    (protocol) -> music.findRemoteTrack(track.getCommunity(), track.toURL()).openDirect(protocol));
+    private static class LocalResult extends Result {
+        private LocalResult(Track track, ResultPriority priority) {
+            super(track.getCommunity(), track.toURL(), track.getUUID(), priority, community -> track);
         }
-
-        public LocalResult(Track track,
-                           ResultPriority priority,
-                           ThrowingFunction<AudioProtocol, AudioProvider, IOException> open) {
-            super(track.getCommunity(),
-                    track.toURL(),
-                    priority,
-                    community -> track,
-                    open);
-        }
-
-        @Override
-        public String getFormat() {
-            try {
-                return get().getFormat();
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        }
-
+    
         @Override
         public InputStream openConnection() throws IOException {
             return get().openRead();
