@@ -1,17 +1,21 @@
 package io.manebot.plugin.music.source;
 
-import com.github.manevolent.ffmpeg4j.FFmpegException;
-import com.github.manevolent.ffmpeg4j.FFmpegIO;
-import com.github.manevolent.ffmpeg4j.FFmpegInput;
+import com.github.manevolent.ffmpeg4j.*;
+import com.github.manevolent.ffmpeg4j.AudioFormat;
+import com.github.manevolent.ffmpeg4j.output.*;
+import com.github.manevolent.ffmpeg4j.stream.output.*;
 import com.github.manevolent.ffmpeg4j.stream.source.FFmpegSourceStream;
 import io.manebot.plugin.audio.mixer.input.AudioProvider;
 import io.manebot.plugin.audio.mixer.input.FFmpegAudioProvider;
+import io.manebot.plugin.audio.mixer.output.*;
+import io.manebot.plugin.audio.resample.*;
+import io.manebot.plugin.music.config.*;
 import org.bytedeco.javacpp.avformat;
 
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 
-import java.util.Map;
+import java.net.*;
+import java.util.*;
 
 public class FFmpegAudioProtocol implements AudioProtocol {
     private final double bufferSeconds;
@@ -20,7 +24,7 @@ public class FFmpegAudioProtocol implements AudioProtocol {
         this.bufferSeconds = bufferSeconds;
     }
 
-    private AudioProvider open(FFmpegIO ioState, String format) throws FFmpegException {
+    private AudioProvider openProvider(FFmpegIO ioState, String format) throws FFmpegException {
         // wrap the IOState.  this would call avformat_alloc_context
         FFmpegInput input = new FFmpegInput(ioState);
 
@@ -58,19 +62,112 @@ public class FFmpegAudioProtocol implements AudioProtocol {
     }
 
     @Override
-    public AudioProvider open(String url, String format, Map<String, String> headers /*TODO ignored*/)
-            throws IOException {
+    public AudioProvider openProvider(URI uri, String format, Map<String, String> headers, int bufferSize) throws IOException {
         try {
-            return open(FFmpegIO.openNativeUrlInput(url), format);
+            return openProvider(FFmpegIO.openNativeUrlInput(uri.toASCIIString()), format);
         } catch (FFmpegException e) {
             throw new IOException(e);
         }
     }
 
     @Override
-    public AudioProvider open(InputStream inputStream, String format) throws IOException {
+    public AudioProvider openProvider(InputStream inputStream, String format, int bufferSize) throws IOException {
         try {
-            return open(FFmpegIO.openInputStream(inputStream, FFmpegIO.DEFAULT_BUFFER_SIZE), format);
+            return openProvider(FFmpegIO.openInputStream(inputStream, FFmpegIO.DEFAULT_BUFFER_SIZE), format);
+        } catch (FFmpegException e) {
+            throw new IOException(e);
+        }
+    }
+    
+    @Override
+    public AudioProvider openProvider(InputStream inputStream) throws IOException {
+        return openProvider(inputStream, FFmpegIO.DEFAULT_BUFFER_SIZE);
+    }
+    
+    @Override
+    public AudioConsumer openConsumer(OutputStream outputStream, AudioDownloadFormat format) throws IOException {
+        AudioFormat audioFormat;
+        try {
+            audioFormat = new AudioFormat(
+                            format.getAudioFormat().getSampleRate(),
+                            format.getAudioFormat().getChannels(),
+                            FFmpeg.guessFFMpegChannelLayout(format.getAudioFormat().getChannels())
+            );
+        } catch (FFmpegException e) {
+            throw new IOException("Problem constructing audio format", e);
+        }
+        
+        FFmpegIO io;
+    
+        try {
+            io = FFmpegIO.openOutputStream(outputStream, FFmpegIO.DEFAULT_BUFFER_SIZE);
+        } catch (FFmpegException e) {
+            throw new IOException("Problem opening FFmpeg output", e);
+        }
+        
+        FFmpegTargetStream targetStream;
+        try {
+           targetStream = new FFmpegTargetStream(format.getContainerFormat(), io, new FFmpegTargetStream.FFmpegNativeOutput());
+        } catch (FFmpegException e) {
+            try {
+                io.close();
+            } catch (Exception ex) {
+                e.addSuppressed(ex);
+            }
+            
+            throw new IOException("Problem opening FFmpeg target stream", e);
+        }
+    
+        AudioTargetSubstream substream;
+        try {
+            substream = targetStream.registerAudioSubstream(format.getAudioCodec(), audioFormat,
+                            Collections.singletonMap("b", Integer.toString(format.getAudioBitrate())));
+        } catch (FFmpegException ex) {
+            try {
+                targetStream.close();
+            } catch (Exception e) {
+                ex.addSuppressed(e);
+            }
+    
+            throw new IOException("Problem opening audio substream", ex);
+        }
+    
+        try {
+            targetStream.writeHeader();
+        } catch (Exception ex) {
+            try {
+                targetStream.close();
+            } catch (Exception e) {
+                ex.addSuppressed(e);
+            }
+        
+            throw new IOException("Problem writing stream header", ex);
+        }
+    
+        return new AudioConsumer() {
+            @Override
+            public void write(float[] buffer, int len) {
+                double time = (double)len / ((double)format.getAudioFormat().getSampleRate() * (double)format.getAudioFormat().getChannels());
+                
+                try {
+                    substream.write(new AudioFrame(0D, 0D, time, buffer, len, audioFormat));
+                } catch (IOException e) {
+                    throw new RuntimeException("Problem writing audio frame", e);
+                }
+            }
+    
+            @Override
+            public void close() throws Exception {
+                targetStream.close();
+            }
+        };
+    }
+    
+    @Override
+    public Resampler openResampler(io.manebot.plugin.music.config.AudioFormat input, io.manebot.plugin.music.config.AudioFormat output, int bufferSize)
+                    throws IOException {
+        try {
+            return new FFmpegResampler(input.toJava(), output.toJava(), bufferSize);
         } catch (FFmpegException e) {
             throw new IOException(e);
         }
